move to separate hooks file to fix tests
