+++ frontend/src/pages/Alerts/ErrorAlert/ErrorAlertPage.tsx
@@ -39,8 +39,8 @@ import { StringParam, useQueryParam } from 'use-query-params'
 
 import LoadingBox from '@/components/LoadingBox'
 import { SearchContext } from '@/components/Search/SearchContext'
+import { useRetentionPresets } from '@/components/Search/SearchForm/hooks'
 import { Search } from '@/components/Search/SearchForm/SearchForm'
-import { useRetentionPresets } from '@/components/Search/SearchForm/utils'
 import { namedOperations } from '@/graph/generated/operations'
 import { ProductType } from '@/graph/generated/schemas'
 import { useSearchTime } from '@/hooks/useSearchTime'
