+++ frontend/src/pages/ErrorsV2/ErrorsV2.tsx
@@ -54,7 +54,7 @@ import {
 
 import { DEMO_PROJECT_ID } from '@/components/DemoWorkspaceButton/DemoWorkspaceButton'
 import { SearchContext } from '@/components/Search/SearchContext'
-import { useRetentionPresets } from '@/components/Search/SearchForm/utils'
+import { useRetentionPresets } from '@/components/Search/SearchForm/hooks'
 import { START_PAGE } from '@/components/SearchPagination/SearchPagination'
 import { GetErrorGroupQuery } from '@/graph/generated/operations'
 import {
