+++ frontend/src/pages/ErrorsV2/SearchPanel/SearchPanel.tsx
@@ -15,8 +15,8 @@ import clsx from 'clsx'
 import { useEffect, useState } from 'react'
 
 import { useSearchContext } from '@/components/Search/SearchContext'
+import { useRetentionPresets } from '@/components/Search/SearchForm/hooks'
 import { SearchForm } from '@/components/Search/SearchForm/SearchForm'
-import { useRetentionPresets } from '@/components/Search/SearchForm/utils'
 import { ErrorFeedCard } from '@/pages/ErrorsV2/ErrorFeedCard/ErrorFeedCard'
 import { useErrorPageNavigation } from '@/pages/ErrorsV2/ErrorsV2'
 import { OverageCard } from '@/pages/Sessions/SessionsFeedV3/OverageCard/OverageCard'
